Address review: split catalog builders into selectors/*.catalog.ts
- Move buildAttributeCatalog/buildDamageCatalog/buildModifiersCatalog
  (the per-namespace catalog factories) into their own files under
  src/lib/stores/runner/selectors/, one per namespace
- Inline the top-level catalog assembly directly into useRunnerSelector
  instead of a separate buildRunnerSelectorCatalog function
- Drop the two now-redundant inline comments
- Add a CONTEXT.md glossary term for Selector Catalog
