+++ src/lib/stores/runner/runnerSelector.ts
@@ -1,51 +1,12 @@
-import { selectTrackWoundModifier } from "#/components/system/damage/damageUtils.ts"
-import type { AttributesContextValue } from "#/lib/contexts/runner/attributesProvider.tsx"
 import { useAttributesContext } from "#/lib/contexts/runner/attributesProvider.tsx"
-import type { AttributeInfo } from "#/system/attributeInfo.ts"
-import type { AttributeKey } from "#/system/attributeKey.ts"
-import { DamageTrackKey } from "#/system/damageTrackKey.ts"
-import type { RunnerData } from "#/system/runnerData.ts"
 
-import { selectMatrixTrack, selectPhysicalTrack, selectStunTrack } from "./damage/damageSlice.selectors.ts"
 import { useRunnerStoreSelector } from "./runnerStore.selectors.ts"
-
-export interface AttributeFacets {
-  baseValue: number
-  info: AttributeInfo
-}
-
-export interface RunnerAttributeCatalog {
-  (key: AttributeKey): AttributeFacets
-  infos: Record<AttributeKey, AttributeInfo>
-}
-
-export interface DamageTrackFacets {
-  current: number
-  max: number
-}
-
-export interface RunnerDamageCatalog {
-  (track: DamageTrackKey): DamageTrackFacets
-  woundMod: number
-}
-
-/**
- * A modifier reachable from more than one catalog namespace — proof-of-concept for the "one
- * implementation, many namespaces" aliasing rule in
- * `docs/adr/0013-unify-runner-state-access.md`. A full `Modifier` catalog (covering every
- * GameEffect-driven stacking modifier, not just wound) is out of scope for this slice.
- */
-export enum Modifier {
-  woundMod = "woundMod",
-}
-
-export interface ModifierFacets {
-  value: number
-}
-
-export interface RunnerModifiersCatalog {
-  (modifier: Modifier): ModifierFacets
-}
+import type { RunnerAttributeCatalog } from "./selectors/attribute.catalog.ts"
+import { buildAttributeCatalog } from "./selectors/attribute.catalog.ts"
+import type { RunnerDamageCatalog } from "./selectors/damage.catalog.ts"
+import { buildDamageCatalog } from "./selectors/damage.catalog.ts"
+import type { RunnerModifiersCatalog } from "./selectors/modifiers.catalog.ts"
+import { buildModifiersCatalog } from "./selectors/modifiers.catalog.ts"
 
 /**
  * The namespaced menu `useRunnerSelector`'s callback picks from. Mirrors the existing
@@ -56,57 +17,6 @@ export interface RunnerSelectorCatalog {
   attribute: RunnerAttributeCatalog
   damage: RunnerDamageCatalog
   modifiers: RunnerModifiersCatalog
-}
-
-const damageTrackSelectors = {
-  [DamageTrackKey.physical]: selectPhysicalTrack,
-  [DamageTrackKey.stun]: selectStunTrack,
-  [DamageTrackKey.matrix]: selectMatrixTrack,
-}
-
-function selectWoundMod(state: RunnerData): number {
-  return selectTrackWoundModifier(DamageTrackKey.physical)(state)
-    + selectTrackWoundModifier(DamageTrackKey.stun)(state)
-}
-
-function buildAttributeCatalog(attributesContext: AttributesContextValue): RunnerAttributeCatalog {
-  const catalog = (key: AttributeKey): AttributeFacets => ({
-    baseValue: attributesContext.values[key] ?? 0,
-    info: attributesContext.infos[key],
-  })
-
-  return Object.assign(catalog, { infos: attributesContext.infos })
-}
-
-function buildDamageCatalog(state: RunnerData): RunnerDamageCatalog {
-  const catalog = (track: DamageTrackKey): DamageTrackFacets => {
-    const { current, max } = damageTrackSelectors[track](state)
-    return { current, max }
-  }
-
-  return Object.assign(catalog, { woundMod: selectWoundMod(state) })
-}
-
-function buildModifiersCatalog(state: RunnerData): RunnerModifiersCatalog {
-  return (modifier: Modifier): ModifierFacets => {
-    switch (modifier) {
-      case Modifier.woundMod:
-        // Same computation as `damage.woundMod` — reachable here too because "wound modifier"
-        // is as much a Modifier concept as a Damage one. Never a second implementation.
-        return { value: selectWoundMod(state) }
-    }
-  }
-}
-
-function buildRunnerSelectorCatalog(
-  state: RunnerData,
-  attributesContext: AttributesContextValue,
-): RunnerSelectorCatalog {
-  return {
-    attribute: buildAttributeCatalog(attributesContext),
-    damage: buildDamageCatalog(state),
-    modifiers: buildModifiersCatalog(state),
-  }
 }
 
 /**
@@ -128,13 +38,14 @@ export function useRunnerSelector<T>(
   picker: (catalog: RunnerSelectorCatalog) => T,
   compare?: (prev: T, next: T) => boolean,
 ): T {
-  // Rules of Hooks — this runs unconditionally on every call, whether or not `picker` ends up
-  // touching `attribute`, so every namespace the catalog can ever expose has to be gathered up
-  // front, not lazily per-namespace.
   const attributesContext = useAttributesContext()
 
   return useRunnerStoreSelector(
-    (state) => picker(buildRunnerSelectorCatalog(state, attributesContext)),
+    (state) => picker({
+      attribute: buildAttributeCatalog(attributesContext),
+      damage: buildDamageCatalog(state),
+      modifiers: buildModifiersCatalog(state),
+    }),
     compare,
   )
 }
